+++ lib/msal-common/src/telemetry/performance/PerformanceEvent.ts
@@ -103,7 +103,23 @@ export enum PerformanceEvents {
     /**
      * Time spent creating default headers for requests to token endpoint
      */
-    BaseClientCreateTokenRequestHeaders = "baseClientCreateTokenRequestHeaders"
+    BaseClientCreateTokenRequestHeaders = "baseClientCreateTokenRequestHeaders",
+
+    /**
+     * Used to measure the time taken for completing embedded-broker handshake (PW-Broker).
+     */
+    BrokerHandhshake = "brokerHandshake",
+
+    /**
+     * acquireTokenByRefreshToken API in BrokerClientApplication (PW-Broker) .
+     */
+    AcquireTokenByRefreshTokenInBroker = "acquireTokenByRefreshTokenInBroker",
+
+    /**
+     * acquireToken API in BrokerClientApplication.
+     * Used to acquire a token on behalf of the embedded application (PW-Broker).
+     */
+    AcquireTokenByBroker = "acquireTokenByBroker"
 }
 
 /**
@@ -239,6 +255,22 @@ export type PerformanceEvent = {
      * @type {string}
      */
     libraryVersion: string,
+
+    /**
+     * Size of the id token
+     *
+     * @type {number}
+     */
+    idTokenSize?: number,
+
+    /**
+     * 
+     * Size of the access token
+     *
+     * @type {number}
+     */
+
+    accessTokenSize?: number,
 
     /**
      * Application name as specified by the app.